Explicitly encoding to UTF-8 on Linux (fixes #11)
CRAN checks equivalent to `rhub::check(platform="debian-clang-devel")`
accuses failure on a couple of test units. This commit fixes it by
making sure, through `iconv()`, that the output is UTF-8 on Linux
systems.

If this problem arises again, perhaps new rules can be made for other
values of `operating.system` ("Windows" and "Darwin").